Re-work master management of egress IP
With new local gateway mode we don't need any specific setup
based on the mode we're running in. Thus ovnkube-master does not
need to differentiate the two. This commit thus does two things:

1. Removes all egressip_local* files
2. Moves everything egressip_shared* -> egressip*
