+++ go-controller/pkg/ovn/egressip.go
@@ -361,6 +361,120 @@ type egressIPMode struct {
 	gatewayIPCache sync.Map
 }
 
+func (e *egressIPMode) addPodEgressIP(eIP *egressipv1.EgressIP, pod *kapi.Pod) error {
+	podIPs := e.getPodIPs(pod)
+	if podIPs == nil {
+		e.podRetry.Store(getPodKey(pod), true)
+		return nil
+	}
+	if e.needsRetry(pod) {
+		e.podRetry.Delete(getPodKey(pod))
+	}
+	for _, status := range eIP.Status.Items {
+		if err := e.createEgressPolicy(podIPs, status, eIP.Name); err != nil {
+			return fmt.Errorf("unable to create logical router policy for status: %v, err: %v", status, err)
+		}
+		if err := createNATRule(podIPs, status, eIP.Name); err != nil {
+			return fmt.Errorf("unable to create NAT rule for status: %v, err: %v", status, err)
+		}
+	}
+	return nil
+}
+
+func (e *egressIPMode) deletePodEgressIP(eIP *egressipv1.EgressIP, pod *kapi.Pod) error {
+	podIPs := e.getPodIPs(pod)
+	if podIPs == nil {
+		return nil
+	}
+	for _, status := range eIP.Status.Items {
+		if err := e.deleteEgressPolicy(podIPs, status, eIP.Name); err != nil {
+			return fmt.Errorf("unable to delete logical router policy for status: %v, err: %v", status, err)
+		}
+		if err := deleteNATRule(podIPs, status, eIP.Name); err != nil {
+			return fmt.Errorf("unable to delete NAT rule for status: %v, err: %v", status, err)
+		}
+	}
+	return nil
+}
+
+func createNATRule(podIPs []net.IP, status egressipv1.EgressIPStatusItem, egressIPName string) error {
+	for _, podIP := range podIPs {
+		if (utilnet.IsIPv6String(status.EgressIP) && utilnet.IsIPv6(podIP)) || (!utilnet.IsIPv6String(status.EgressIP) && !utilnet.IsIPv6(podIP)) {
+			natIDs, err := findNatIDs(egressIPName, podIP.String(), status.EgressIP)
+			if err != nil {
+				return err
+			}
+			if natIDs == nil {
+				_, stderr, err := util.RunOVNNbctl(
+					"--id=@nat",
+					"create",
+					"nat",
+					"type=snat",
+					fmt.Sprintf("logical_port=k8s-%s", status.Node),
+					fmt.Sprintf("external_ip=%s", status.EgressIP),
+					fmt.Sprintf("logical_ip=%s", podIP),
+					fmt.Sprintf("external_ids:name=%s", egressIPName),
+					"--",
+					"add",
+					"logical_router",
+					fmt.Sprintf("GR_%s", status.Node),
+					"nat",
+					"@nat",
+				)
+				if err != nil {
+					return fmt.Errorf("unable to create nat rule, stderr: %s, err: %v", stderr, err)
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func deleteNATRule(podIPs []net.IP, status egressipv1.EgressIPStatusItem, egressIPName string) error {
+	for _, podIP := range podIPs {
+		if (utilnet.IsIPv6String(status.EgressIP) && utilnet.IsIPv6(podIP)) || (!utilnet.IsIPv6String(status.EgressIP) && !utilnet.IsIPv6(podIP)) {
+			natIDs, err := findNatIDs(egressIPName, podIP.String(), status.EgressIP)
+			if err != nil {
+				return err
+			}
+			for _, natID := range natIDs {
+				_, stderr, err := util.RunOVNNbctl(
+					"remove",
+					"logical_router",
+					fmt.Sprintf("GR_%s", status.Node),
+					"nat",
+					natID,
+				)
+				if err != nil {
+					return fmt.Errorf("unable to remove nat from logical_router, stderr: %s, err: %v", stderr, err)
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func findNatIDs(egressIPName, podIP, egressIP string) ([]string, error) {
+	natIDs, stderr, err := util.RunOVNNbctl(
+		"--format=csv",
+		"--data=bare",
+		"--no-heading",
+		"--columns=_uuid",
+		"find",
+		"nat",
+		fmt.Sprintf("external_ids:name=%s", egressIPName),
+		fmt.Sprintf("logical_ip=%s", podIP),
+		fmt.Sprintf("external_ip=%s", egressIP),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("unable to find nat ID, stderr: %s, err: %v", stderr, err)
+	}
+	if natIDs == "" {
+		return nil, nil
+	}
+	return strings.Split(natIDs, "\n"), nil
+}
+
 func (e *egressIPMode) getGatewayRouterJoinIP(node string, wantsIPv6 bool) (net.IP, error) {
 	var gatewayIPs []net.IP
 	if item, exists := e.gatewayIPCache.Load(node); exists {
@@ -407,10 +521,8 @@ func (e *egressIPMode) needsRetry(pod *kapi.Pod) bool {
 }
 
 // createEgressPolicy uses logical router policies to force egress traffic to the egress node, for that we need
-// to retrive the internal gateway router IP attached to the egress node. This method handles both the shared and
-// local gateway mode case, and in the local case (i.e: packetMark != 0) uses the optional packetMark field
-// on the logical router policy to properly L3 SNAT on the node side.
-func (e *egressIPMode) createEgressPolicy(podIps []net.IP, status egressipv1.EgressIPStatusItem, packetMark uint32, egressIPName string) error {
+// to retrive the internal gateway router IP attached to the egress node.
+func (e *egressIPMode) createEgressPolicy(podIps []net.IP, status egressipv1.EgressIPStatusItem, egressIPName string) error {
 	isEgressIPv6 := utilnet.IsIPv6String(status.EgressIP)
 	gatewayRouterIP, err := e.getGatewayRouterJoinIP(status.Node, isEgressIPv6)
 	if err != nil {
@@ -424,47 +536,27 @@ func (e *egressIPMode) createEgressPolicy(podIps []net.IP, status egressipv1.Egr
 		} else if !isEgressIPv6 && !utilnet.IsIPv6(podIP) {
 			filterOption = fmt.Sprintf("ip4.src == %s", podIP.String())
 		}
-		policyIDs, err := findLRPolicyIDs(filterOption, egressIPName, gatewayRouterIP)
+		policyIDs, err := findReroutePolicyIDs(filterOption, egressIPName, gatewayRouterIP)
 		if err != nil {
 			return err
 		}
 		if policyIDs == nil {
-			if packetMark != 0 {
-				_, stderr, err = util.RunOVNNbctl(
-					"--id=@lr-policy",
-					"create",
-					"logical_router_policy",
-					"action=reroute",
-					fmt.Sprintf("match=\"%s\"", filterOption),
-					fmt.Sprintf("priority=%v", egressIPReroutePriority),
-					fmt.Sprintf("nexthop=%s", gatewayRouterIP),
-					fmt.Sprintf("external_ids:name=%s", egressIPName),
-					fmt.Sprintf("options:pkt_mark=%v", packetMark),
-					"--",
-					"add",
-					"logical_router",
-					ovnClusterRouter,
-					"policies",
-					"@lr-policy",
-				)
-			} else {
-				_, stderr, err = util.RunOVNNbctl(
-					"--id=@lr-policy",
-					"create",
-					"logical_router_policy",
-					"action=reroute",
-					fmt.Sprintf("match=\"%s\"", filterOption),
-					fmt.Sprintf("priority=%v", egressIPReroutePriority),
-					fmt.Sprintf("nexthop=%s", gatewayRouterIP),
-					fmt.Sprintf("external_ids:name=%s", egressIPName),
-					"--",
-					"add",
-					"logical_router",
-					ovnClusterRouter,
-					"policies",
-					"@lr-policy",
-				)
-			}
+			_, stderr, err = util.RunOVNNbctl(
+				"--id=@lr-policy",
+				"create",
+				"logical_router_policy",
+				"action=reroute",
+				fmt.Sprintf("match=\"%s\"", filterOption),
+				fmt.Sprintf("priority=%v", egressIPReroutePriority),
+				fmt.Sprintf("nexthop=%s", gatewayRouterIP),
+				fmt.Sprintf("external_ids:name=%s", egressIPName),
+				"--",
+				"add",
+				"logical_router",
+				ovnClusterRouter,
+				"policies",
+				"@lr-policy",
+			)
 			if err != nil {
 				return fmt.Errorf("unable to create logical router policy: %s, stderr: %s, err: %v", status.EgressIP, stderr, err)
 			}
@@ -486,7 +578,7 @@ func (e *egressIPMode) deleteEgressPolicy(podIps []net.IP, status egressipv1.Egr
 		} else if !utilnet.IsIPv6(podIP) && !utilnet.IsIPv6String(status.EgressIP) {
 			filterOption = fmt.Sprintf("ip4.src == %s", podIP.String())
 		}
-		policyIDs, err := findLRPolicyIDs(filterOption, egressIPName, gatewayRouterIP)
+		policyIDs, err := findReroutePolicyIDs(filterOption, egressIPName, gatewayRouterIP)
 		if err != nil {
 			return err
 		}
@@ -506,7 +598,7 @@ func (e *egressIPMode) deleteEgressPolicy(podIps []net.IP, status egressipv1.Egr
 	return nil
 }
 
-func findLRPolicyIDs(filterOption, egressIPName string, gatewayRouterIP net.IP) ([]string, error) {
+func findReroutePolicyIDs(filterOption, egressIPName string, gatewayRouterIP net.IP) ([]string, error) {
 	policyIDs, stderr, err := util.RunOVNNbctl(
 		"--format=csv",
 		"--data=bare",
